Update how sim code is added
